Moved delta bar ui to separate file.
